requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
p==0.0.0
pi==0.1.2
requests==2.22.0
urllib3==1.25.2
Werkzeug==0.15.4
